+++ components/pages/heic.tsx
